+++ lib/query_owl/configuration.rb
@@ -5,7 +5,7 @@ module QueryOwl
 
     attr_reader :log_level, :backtrace_filter
     attr_accessor :enabled, :slow_query_threshold_ms, :n_plus_one_threshold, :backtrace_lines,
-                  :raise_on_n_plus_one
+                  :raise_on_n_plus_one, :event_store_size
 
     def initialize
       @enabled                 = Rails.env.development?
@@ -15,6 +15,7 @@ module QueryOwl
       @backtrace_lines         = 5
       @backtrace_filter        = DEFAULT_BACKTRACE_FILTER
       @raise_on_n_plus_one     = false
+      @event_store_size        = 100
     end
 
     def log_level=(level)